+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## 3.6.8
 * `NEW` command `lua.exportDocument` . VSCode will display this command in the right-click menu
+* `CHG` setting `Lua.workspace.supportScheme` has been removed. All schemes are supported if the language id is `lua`
 * `FIX` [#1831]
 * `FIX` [#1838]
 * `FIX` [#1841]